+++ kit/structs.go
@@ -77,33 +77,10 @@ type Team struct {
 	Bid						int								`json:"bid"`
 }
 
-// ------------------------------------------------------------------------------------------------
-// Thanks to MMJ -- aka themmj on Github -- for help with this...
-
 type Pos struct {
 	X						int
 	Y						int
 }
-
-func (p *Pos) UnmarshalJSON(data []byte) error {
-	var v [2]int
-	err := json.Unmarshal(data, &v)
-	if err != nil {
-		return err
-	}
-	p.X = v[0]
-	p.Y = v[1]
-	return nil
-}
-
-func (p Pos) MarshalJSON() ([]byte, error) {
-	var v [2]int
-	v[0] = p.X
-	v[1] = p.Y
-	return json.Marshal(&v)
-}
-
-// ------------------------------------------------------------------------------------------------
 
 type Action struct {
 	Type					ActionType
@@ -118,91 +95,3 @@ type ActionType int
 type Direction int
 type Resource int
 type FactoryActionType int
-
-func (a *Action) UnmarshalJSON(data []byte) error {
-	var v [6]int
-	err := json.Unmarshal(data, &v)
-	if err != nil {
-		return err
-	}
-	a.Type = ActionType(v[0])
-	a.Direction = Direction(v[1])
-	a.Resource = Resource(v[2])
-	a.Amount = v[3]
-	a.Recycle = v[4]
-	a.N = v[5]
-	return nil
-}
-
-func (a Action) MarshalJSON() ([]byte, error) {
-	var v [6]int
-	v[0] = int(a.Type)
-	v[1] = int(a.Direction)
-	v[2] = int(a.Resource)
-	v[3] = a.Amount
-	v[4] = a.Recycle
-	v[5] = a.N
-	return json.Marshal(&v)
-}
-
-// ------------------------------------------------------------------------------------------------
-// Our Factory and Unit types have some embedded structs for user convenience, but this seems to
-// mean (as far as I can tell) that we need custom unmarshalling for them... (?)
-
-type unit_tmp struct {
-	TeamId					int								`json:"team_id"`
-	UnitId					string							`json:"unit_id"`
-	Power					int								`json:"power"`
-	UnitType				string							`json:"unit_type"`
-	Pos						Pos								`json:"pos"`
-	Cargo					Cargo							`json:"cargo"`
-	ActionQueue				[]Action						`json:"action_queue"`
-}
-
-func (u *Unit) UnmarshalJSON(data []byte) error {
-
-	var v unit_tmp
-	err := json.Unmarshal(data, &v)
-	if err != nil {
-		return err
-	}
-
-	u.TeamId = v.TeamId
-	u.UnitId = v.UnitId
-	u.Power = v.Power
-	u.UnitType = v.UnitType
-	u.Pos = v.Pos
-	u.Cargo = v.Cargo
-	u.ActionQueue = v.ActionQueue
-
-	return nil
-}
-
-// ------------------------------------------------------------------------------------------------
-
-type factory_tmp struct {
-	TeamId					int								`json:"team_id"`
-	UnitId					string							`json:"unit_id"`
-	Power					int								`json:"power"`
-	Pos						Pos								`json:"pos"`
-	Cargo					Cargo							`json:"cargo"`
-	StrainId				int								`json:"strain_id"`
-}
-
-func (fc *Factory) UnmarshalJSON(data []byte) error {
-
-	var v factory_tmp
-	err := json.Unmarshal(data, &v)
-	if err != nil {
-		return err
-	}
-
-	fc.TeamId = v.TeamId
-	fc.UnitId = v.UnitId
-	fc.Power = v.Power
-	fc.Pos = v.Pos
-	fc.Cargo = v.Cargo
-	fc.StrainId = v.StrainId
-
-	return nil
-}
